requirements.txt
anyio==4.4.0
APScheduler==3.10.4
certifi==2024.7.4
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.3
pytz==2024.1
six==1.16.0
sniffio==1.3.1
tzdata==2024.1
tzlocal==5.2
